refactor : groupService 수정

+++ src/main/java/com/example/Waffle/repository/UserGroupRepository.java
@@ -15,5 +15,5 @@ public interface UserGroupRepository extends JpaRepository<UserGroupEntity, User
 
     Optional<UserGroupEntity> findByUserIdAndGroupId(Long userId, Long groupId);
 
-    Optional<UserGroupEntity> deleteByGroupId(int groupId);
+    Optional<UserGroupEntity> deleteByGroupId(Long groupId);
 }
